Creation of the system to make the interactions and change status of a person's vital model

+++ core/models/personVitalModel.h
@@ -25,7 +25,7 @@ typedef struct Vital_Model {
     int maxDurationOfInfectionInDays;
 
     /* Taxa de uma pessoa ficar imune após recuperação */
-    int immunityRate;
+    float immunityRate;
 
 } Vital_Model;
